Bug fix for crasher in `test_diff_commonOverlap` when running tests with Swift Package manager.
A litte context: I'm creating a SPM package, using SPM to build the objective part and translating the tests to swift. While doing so I found that
```
  XCTAssertEqual((NSUInteger)3, [dmp diff_commonOverlapOfFirstString:@"abc" andSecondString:@"abcd"], @"Detect any suffix/prefix overlap. Whole case.");
```
crashes with
```
fatal error: subscript: subrange start precedes String start
```

It looks like this is actually a genuine bug in `diff_commonOverlap` where it does a pointer comparison instead of a string comparison.